+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Feature
 
+- Added 'checkAndNormalizeUrl' function in URLUtils. @giuliaghisini
+
 ### Bugfix
 
 - Used UniversalLink and PreviewImage components where needed, to right handle link and images. @giuliaghisini